preparando o jogo

+++ src/main/kotlin/models/Tile.kt
@@ -1,6 +1,6 @@
 package models
 
-data class Tile(private val oneNumber: Byte, private val twoNumber: Byte) {
+data class Tile(private val oneNumber: Int, private val twoNumber: Int) {
 
     val isDoble = oneNumber == twoNumber;
